CODAGE/DEBUG: ajout/delete refresh lastestsong

+++ frontend/src/components/music/small/song/SmallSong.jsx
@@ -3,8 +3,15 @@ import { Link, useNavigate } from "react-router-dom";
 import { AuthContext } from "../../../../context/auth-context";
 import ModalMessageErreur from "../../../UIElements/ModalMessageErreur";
 
-export default function SmallSong({ title, artist, filetype, location, id }) {
-  const { user } = useContext(AuthContext);
+export default function SmallSong({
+  title,
+  artist,
+  filetype,
+  location,
+  id,
+  onUpdate,
+}) {
+  const { user, token } = useContext(AuthContext);
 
   const musicRole = user?.perms?.music;
 
@@ -12,19 +19,24 @@ export default function SmallSong({ title, artist, filetype, location, id }) {
   const canModnDelete = musicRole === "admin";
 
   const deleteSong = async () => {
+    if (!token) {
+      console.error("No token found");
+      return;
+    }
     try {
       const response = await fetch(
         `http://localhost:5000/api/music/${id}/delete`,
         {
           method: "DELETE",
           headers: {
-            Authorization: "Bearer " + user.token,
+            Authorization: "Bearer " + token,
           },
         },
       );
       const message = await response.json();
       if (response.ok) {
         console.log("The song has been deleted successfully.");
+        onUpdate();
       }
     } catch (err) {
       console.error("Failed to delete song", err);
